Added PATCH

+++ entity/serializers.py
@@ -3,7 +3,14 @@ from rest_framework import serializers
 from .models import Address, Entity, Resource
 
 
+def find_id(validated_address):
+    return len(list(filter(lambda found: 'id' in found, validated_address)))
+
+
+# noinspection PyUnresolvedReferences
 class ResourceSerializer(serializers.ModelSerializer):
+    id = serializers.IntegerField(required=False)
+
     class Meta:
         model = Resource
         fields = (
@@ -15,7 +22,6 @@ class ResourceSerializer(serializers.ModelSerializer):
 
     def to_representation(self, instance):
         rep = super().to_representation(instance)
-        rep.id = instance.pk
         return rep
 
     def create(self, validated_data):
@@ -28,8 +34,10 @@ class ResourceSerializer(serializers.ModelSerializer):
         return instance
 
 
+# noinspection PyUnresolvedReferences
 class AddressSerializer(serializers.ModelSerializer):
     resource = ResourceSerializer(many=True)
+    id = serializers.IntegerField(required=False)
 
     class Meta:
         model = Address
@@ -49,7 +57,6 @@ class AddressSerializer(serializers.ModelSerializer):
 
     def to_representation(self, instance):
         rep = super().to_representation(instance)
-        rep.id = instance.pk
         return rep
 
     def create(self, validated_data):
@@ -62,17 +69,32 @@ class AddressSerializer(serializers.ModelSerializer):
         return instance
 
     def update(self, instance, validated_data):
-        resources = instance.resource.all()
-        resources.all().delete()
-
-        validated_resource = validated_data.pop('resource')
-
-        for resource in validated_resource:
-            instance.resource.create(**resource)
+        if 'resource' in validated_data:
+            validated_resource = validated_data.pop('resource')
+            if len(list(filter(lambda resource: 'id' in resource, validated_resource))):
+                self.merge(validated_resource)
+            else:
+                self.replace(instance, validated_resource)
 
         update_instance(instance, validated_data)
+
         instance.save()
         return instance
+
+    @staticmethod
+    def merge(validated_resource):
+        for item in validated_resource:
+            pk = item.pop('id')
+            resource_instance = Resource.objects.get(pk=pk)
+            update_instance(resource_instance, item)
+            resource_instance.save()
+
+    @staticmethod
+    def replace(instance, validated_resource):
+        resources = instance.resource.all()
+        resources.all().delete()
+        for resource in validated_resource:
+            instance.resource.create(**resource)
 
 
 def update_instance(instance, validated_data):
@@ -83,8 +105,10 @@ def update_instance(instance, validated_data):
             pass
 
 
+# noinspection PyUnresolvedReferences
 class EntitySerializer(serializers.ModelSerializer):
     address = AddressSerializer(many=True)
+    id = serializers.IntegerField(required=False)
 
     class Meta:
         model = Entity
@@ -103,7 +127,6 @@ class EntitySerializer(serializers.ModelSerializer):
 
     def to_representation(self, instance):
         rep = super().to_representation(instance)
-        rep.id = instance.pk
         return rep
 
     def create(self, validated_data):
@@ -123,21 +146,35 @@ class EntitySerializer(serializers.ModelSerializer):
         return instance
 
     def update(self, instance, validated_data):
-        addresses = instance.address.all()
+        if 'address' in validated_data:
+            validated_address = validated_data.pop('address')
+            if not find_id(validated_address):
+                addresses = instance.address.all()
+                for address in addresses:
+                    address.resource.all().delete()
+                addresses.all().delete()
 
-        for address in addresses:
-            address.resource.all().delete()
-        addresses.all().delete()
+                for address in validated_address:
+                    # extract resources from address
+                    validated_resource = address.pop('resource')
+                    created_address = instance.address.create(**address)
 
-        validated_address = validated_data.pop('address')
+                    for resource in validated_resource:
+                        created_address.resource.create(**resource)
+            else:
+                for address in validated_address:
+                    if 'resource' in address:
+                        validated_resources = address.pop('resource')
+                        for resource in validated_resources:
+                            pk = resource.pop('id')
+                            resource_instance = Resource.objects.get(pk=pk)
+                            update_instance(resource_instance, resource)
+                            resource_instance.save()
 
-        for address in validated_address:
-            # extract resources from address
-            validated_resource = address.pop('resource')
-            created_address = instance.address.create(**address)
-
-            for resource in validated_resource:
-                created_address.resource.create(**resource)
+                    pk = address.pop('id')
+                    address_instance = Address.objects.get(pk=pk)
+                    update_instance(address_instance, address)
+                    address_instance.save()
 
         update_instance(instance, validated_data)
         instance.save()
